requirements.txt
mlx
wand
transformers
Pillow